chore; move functinos
